Change snapshot taking method to pause running VMs
Previously the takeSnapshot call passed false for pause argument,
indicating a "live" snapshot to be taken. That no longer worked with VirtualBox 7.1.
The User's Manual for VirtualBox 7.1 says that running VMs are paused during snapshotting,
which corresponds to the "online" snapshot method. The value is changed to true to align
with the documentation and make snapshots working via the web UI.